refactor into packages of FXML implementation
