+++ src/main/java/duke/userInterface/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.userInterface;
 
 import java.io.IOException;
 import java.util.Collections;
